Object#dup does more than we need and causes problems with Globalize3.
`PaperTrail::Model::InstanceMethods#item_before_change` was using
`Object#dup` to build a _temporary_ representation of the model before any
changes. I'm pretty confident that there is no need to completely
duplicate the model in order to do this and just building a new model
with the same attribute values is sufficient.

The reason I want to make this change is that using `Object#dup` was
breaking integration with Globalize3. This library adds an instance
variable `@globalize` to the model in order to reference an instance of
`Globalize::ActiveRecord::Adapter`. When the model was being duplicated,
the reference to this adapter instance was duplicated and inappropriate
changes were then made to its state. The fix described above means that
a new _temporary_ instance of the adapter is built to go along with the
_temporary_ instance of the model thus avoiding inappropriate changes to
the "real" / "original" instance of the adapter.

With this change, there is no longer any need to explicitly copy
timestamp attributes or the ID, both of which will now be copied by the
generic attribute-copying mechanism.