Use class inst. var. instead of class var.
I use the term "class instance variable" and "class variable"
even though `PaperTrail` is a module. This nomenclature is
conventional in ruby.

> module A; @@x = :x; puts defined?(@@x); end
> #=> "class variable"

Actually, because `PaperTrail` is a module, the use of a class
variable here is not as problematic as in an actual class, so
this change is solely to conform the the Ruby Style Guide.

> Avoid the usage of class (@@) variables due to their "nasty"
> behavior in inheritance.
> https://github.com/bbatsov/ruby-style-guide#no-class-vars

As stated above, because `PaperTrail` is a module, there is no
concern over inheritance, and this change is solely to conform
to the style guide.