Remove my own misleading comment
The `private` keyword does, in fact, make those instance methods
private. I think what I was trying to say is that the host class
(the class that VersionConcern is mixed into) will be able to use
these methods.  They *will* be private after being mixed in, but
they are not "api private", if that makes sense.

[ci skip]