Extract test to new file: ar_id_map_test.rb
This test is quite strange, and has very little explanation.

In an effort to understand it better, move it and its context to a new file.

In particular, reopening the ActiveRecord::IdentityMap module inside a test
is unusual. More importantly, it's not clear why we're testing a method that
we do not implement.