Fix the ar3 build on 4.1-stable branch (#820)
Fixes the travis build of the AR 3 gemfile, which was failing to resolve dependencies under ruby 1.8.7. The changes are:

activerecord = 3.2.22.2 (was ~> 3.2)
i18n < 0.7 (was ~> 0.6.11)
railties > 0 (was ~> 3.0)
sinatra ~> 1.1.4 (was ~> 1.0)
rspec-rails ~> 3.4.2 (was ~> 3.1.0)

These changes are only applicable to PT 4, as PT 5 dropped support for ruby 1.8.7 and also uses the appraisal gem now to manage its various AR gemfiles.