Extract query object: WhereObjectChanges
This is the first query object in PT, so I'll explain the rationale.

In my mind, the Query Object pattern is a subclass of the Method
Object pattern.

http://refactoring.com/catalog/replaceMethodWithMethodObject.html

The goal is simple, to break up a very large method. In this case,
there is another big advantage; it gives us a place to extract
methods that will actually be private. Better still, little private
methods will not pollut the namespace of classes that mix in
`VersionConcern`.