Move sinatra integration to separate gem
Sinatra has been slow to release a version compatible with AR 5,
so having a dev. dependency on sinatra has been painful. This is the
main reason for the move, but there are the other, usual reasons:

- Better organization of github issues
- Independent development. The sinatra integration can target older
  versions of PT, for example.
- Rails users don't want sinatra integration in their apps, even if
  it's basically inactive.

And, of course, there is the inevitable downside that sinatra integration
is now a second-class citizen. That's the trade-off with separate gems.