Improves performance for large datasets by outsourcing subselects, so the subselect only happens once
On a dataset of ~300 000 entries, one query took about 20 seconds, which could be reduced to less than one ms.
This solution may cause problems when the subselect returns a large set of indices, therefore I tested this:
On the same dataset an query "SELECT * FROM versions WHERE ID in (30 000 different ids)" took about 112ms, but it was a direct sql statement so network latency is not included. Also, the subquery which collects the ids is not included. But even if it were, it's still better than 20 seconds.

However, if the subselect returns 30 000 different ids, the versions table may be much larger than 300 000 entries and therefore even slower when using an subselect.