Set @in_after_callback correctly during record_destroy
A destroy event can be handled either in a before-callback
or an after-callback. It's configurable.

However, we were not setting @in_after_callback accordingly.

I'm not aware of a specific bug caused by failing to set
this instance variable correctly, but if the instance variable
is a lie, it's easy to imagine such a bug now or in the future.