Handle migration name timestamp conflict with sleep
We sleep until the next whole second because that is the precision of the
timestamp that rails puts in generator filenames. If we didn't sleep,
there's a good chance two tests would run within the same second and
generate the same exact migration filename. Then, even though we delete the
generated migrations after running them, some form of caching (perhaps
filesystem, perhaps rails) will run the cached migration file.

This implementation, with sleep, replaces a previous implementation
using a "dummy migration" that forced rails to use the next timestamp.
This was a very clever solution, and in the best case was N seconds faster
than the sleep solution, where N is the number of times we run
`generate_and_migrate` in the test suite. Currently N is 3, and I'm willing
to live with 3s of sleep, and I prefer its simplicity.

Perhaps most importantly, this implementation does not require any state
to be persisted across tests (ie. the @migrator ivar in the previous
implementation. Tests should be completely independent of eachother.