Readme doesn't say generator creates initialiser
Fixes #1163

As of 1d42d0d0ab1537123b115e4217badec0ad165a61 the generator does not create an initializer.

This seems to have been unintentional, but it isn't actually required for PaperTrail to run correctly and there are [no plans to reinstate it](https://github.com/paper-trail-gem/paper_trail/issues/1163#issuecomment-435079777) so it's best to remove references to it from the documentation.

Also remove references to this in the generator description