Tests: Fix overly-specific YAML test
We don't care exactly how the YAML is serialized, as long as it parses
into the same thing.

On my machine, for example, ruby 2.7 writes a nil in YAML with trailing
whitespace, and ruby 3.0 does not. But, trailing whitespace doesn't matter
in this case. It still deserializes into a nil.