Query Versions Where Object Changed From Attributes (#1286)
* Add rails-controller-testing

In order to get all of the tests to pass, this adds in and configures
rails-controller-testing so that `assigns`is available in controller and
integration tests.

* Query Versions Where Object Changed From Attributes

This proposes an extension to the public API to provide more targeted
querying of object changes. `where_object_changes` will look for either
side of the change of the attributes provided - either versions where
the attribute changed __from__ the provided value, or changed __to__ the
provided value.

The proposed `where_object_changes_from` addition focuses only on one
side of that equation. If you want to find versions where the
attribute(s) explicitly changed *from* some known value, this will only
show those changes, as opposed to both *from* and *to*.

* Custom Serializer where_object_changes_from implementation

This mirrors the implementation of WhereObjectChanges#text in
WhereObjectChangesFrom#text in the case of a custom serializer. The
provided YAML serializer does not support this method.

* Document New API Method for Custom Adapter

This provides awareness of the new method that custom object changes
adapters may need to implement to support this change.

* remove rails-controller-testing