Query Versions Where Attribute Changed
This adds in a new query method `where_attribute_changes` which will
look for any versions where there was a change to the attribute
specified.

In instances where you need to look for any changes to a particular
attribute, but don't yet know, or don't care, what the value of the
attribute was that changed at all, this is intended to answer that
question.