Try to create a version record when associated object is touched
Before #1285, a version record was created when associated object is touched.
But because of Rails don't track implicit touch mutation[1], now PaperTail
don't create a version record was created when associated object is touched.

This patch try to keep creating a version record in that case by doing checking an
event and changed value.

Fixes #1339.

[1]: https://github.com/rails/rails/commit/dcb825902d79d0f6baba956f7c6ec5767611353e