Fix: n+1 when loading a changeset
When producing a changeset, the item is loaded from the database in
order to check it's class. This came about because of issues that
emerged when supporting STI.

Unfortunately, the result is that the record is retrieved once for every
version despite the fact that the item class will be the same.

The documented reason for moving away from using item_type to know the
class was that it returns the wrong class when working with STI but
there is an item_subtype which returns the correct class.