Change return value of #save_with_version when disabled
This patch changes the return value behavior of #save_with_version to
return a boolean if PaperTrail is disabled.

Previously #save_with_version would return `nil` if version tracking is
disabled. After applying this commit it will return the result of
calling #save on the model.

The main reason for changing this behavior is to be able to react to
failing saves when using #save_with_version in our application - for
example when saving with invalid data.