Add model option for not synchronizing timestamps
Paper trail automatically synchronizes the version
created_at timestamp to be the same as the updated_at
timestamp of the record being updated. This pull-requests
adds a configurable model-level option to disable this
behavior and instead have created_at be populated with
the timestamp when the version was inserted into the database.

The default remains to synchronize version timestamps