Add support for ActiveRecord Encryption
For PostgreSQL json and jsonb columns

When Active Record encryption is used, use Active Record serializer to encrypt values in the object and object_changes column.

NOTE: when nondeterministic encryption is used, the cyphertext will not be the same as that for the record in the database, because it is encrypted separately.  It will decrypt to the same value.

Update CHANGELOG.md

fixes #1392